[frontend] fix local storage in entities list at step 1 of creation
