+++ opencti-platform/opencti-front/src/private/components/common/stix_core_relationships/StixCoreRelationshipCreationFromEntityHeader.tsx
@@ -4,9 +4,6 @@ import { graphql, useFragment } from 'react-relay';
 import {
   StixCoreRelationshipCreationFromEntityHeader_stixCoreObject$key,
 } from '@components/common/stix_core_relationships/__generated__/StixCoreRelationshipCreationFromEntityHeader_stixCoreObject.graphql';
-import {
-  StixCoreRelationshipCreationFromEntityStixCoreObjectsLinesQuery$variables,
-} from '@components/common/stix_core_relationships/__generated__/StixCoreRelationshipCreationFromEntityStixCoreObjectsLinesQuery.graphql';
 import { useFormatter } from '../../../../components/i18n';
 import Drawer from '../drawer/Drawer';
 import { TargetEntity } from './StixCoreRelationshipCreationFromEntity';
@@ -15,8 +12,6 @@ import StixCoreRelationshipCreationSelectEntityStage from './StixCoreRelationshi
 import StixCoreRelationshipCreationFormStage from './StixCoreRelationshipCreationFormStage';
 import { CreateRelationshipContext } from './CreateRelationshipContextProvider';
 import { computeTargetStixCyberObservableTypes, computeTargetStixDomainObjectTypes } from '../../../../utils/stixTypeUtils';
-import { usePaginationLocalStorage } from '../../../../utils/hooks/useLocalStorage';
-import { useBuildEntityTypeBasedFilterContext } from '../../../../utils/filters/filtersUtils';
 import { PaginationOptions } from '../../../../components/list_lines';
 
 /**
@@ -59,6 +54,7 @@ StixCoreRelationshipCreationFromEntityHeaderProps
   // Drawer and form control
   const [open, setOpen] = useState<boolean>(false);
   const [step, setStep] = useState<number>(0);
+  const [searchPaginationOptions, setSearchPaginationOptions] = useState<PaginationOptions>({});
   const handleOpen = () => setOpen(true);
   const handleClose = () => {
     setOpen(false);
@@ -70,25 +66,7 @@ StixCoreRelationshipCreationFromEntityHeaderProps
     setTargetEntities([]);
   };
 
-  // paginationOptions of entities list in first step of relationship creation form
   const storageKey = `stixCoreRelationshipCreationFromEntity-${stixCoreObject.id}-${targetStixDomainObjectTypes.join('-')}-${targetStixCyberObservableTypes.join('-')}`;
-  const { viewStorage, helpers } = usePaginationLocalStorage<StixCoreRelationshipCreationFromEntityStixCoreObjectsLinesQuery$variables>(
-    storageKey,
-    {
-      orderAsc: false,
-      sortBy: '_score',
-      searchTerm: '',
-    },
-    true,
-  );
-  const { searchTerm, orderAsc, sortBy, filters } = viewStorage;
-  const contextFilters = useBuildEntityTypeBasedFilterContext(stixCoreObjectTypes, filters);
-  const searchPaginationOptions: PaginationOptions = {
-    search: searchTerm,
-    filters: contextFilters,
-    orderBy: sortBy,
-    orderMode: orderAsc ? 'asc' : 'desc',
-  } as PaginationOptions;
 
   return (
     <>
@@ -129,9 +107,7 @@ StixCoreRelationshipCreationFromEntityHeaderProps
               setTargetEntities={setTargetEntities}
               virtualEntityTypes={stixCoreObjectTypes}
               handleClose={handleClose}
-              searchPaginationOptions={searchPaginationOptions}
-              contextFilters={contextFilters}
-              helpers={helpers}
+              setSearchPaginationOptions={setSearchPaginationOptions}
             />
           ) : (
             <StixCoreRelationshipCreationFormStage
